Fix: (FE) fix page component ( for metadata )
모든 페이지 메타데이터 적용을 위해 코드 수정
( "use client" 제거 및 컴포넌트 분리 )
