+++ src/@types/environment.d.ts
@@ -1,5 +1,7 @@
 namespace NodeJS {
   interface ProcessEnv extends NodeJS.ProcessEnv {
+    BASE_URL: string;
+
     DATABASE_URL: string;
 
     ACCESS_SECRET: string;
