+++ src/apis/index.ts
@@ -7,6 +7,14 @@ export const serverInstance = axios.create({
   timeout: 4000,
 });
 
+/** 2023/04/30 - SSR에 사용하는 axios 인스턴스 - by 1-blue */
+export const ssrInstance = axios.create({
+  baseURL: process.env.BASE_URL + "/api",
+  withCredentials: true,
+  timeout: 1000,
+  headers: { "Cache-Control": "no-cache" },
+});
+
 export * from "./auth";
 export * from "./me";
 export * from "./user";
