+++ src/app/loading.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+// component
+import Spinner from "@src/components/common/Spinner";
+
+/** 2023/04/30 - 전체 로딩 컴포넌트 - by 1-blue */
+const LoadingPage = () => <Spinner.Page />;
+
+export default LoadingPage;
