+++ src/app/not-found.tsx
@@ -3,4 +3,7 @@
 // component
 import NotFound from "@src/components/common/NotFound";
 
-export default NotFound;
+/** 2023/04/30 - 잘못된 경로 페이지 - by 1-blue */
+const NotFoundPage = () => <NotFound />;
+
+export default NotFoundPage;
