+++ src/components/common/ThemeToggleButton/index.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { useContext } from "react";
 import { useTheme } from "styled-components";
