+++ src/hooks/query/usePosts.ts
@@ -1,4 +1,4 @@
-import { useInfiniteQuery } from "react-query";
+import { type InfiniteData, useInfiniteQuery } from "react-query";
 
 // api
 import { apiServicePosts } from "@src/apis";
@@ -11,10 +11,11 @@ import type { ApiFetchPostsResponse } from "@src/types/api";
 interface Props {
   take: number;
   lastIdx?: number;
+  initialData?: InfiniteData<ApiFetchPostsResponse>;
 }
 
 /** 2023/04/08 - 게시글들을 얻는 훅 - by 1-blue ( 2023/04/10 ) */
-const usePosts = ({ take, lastIdx = -1 }: Props) => {
+const usePosts = ({ take, lastIdx = -1, initialData }: Props) => {
   const { data, fetchNextPage, hasNextPage, isFetching } =
     useInfiniteQuery<ApiFetchPostsResponse>(
       [queryKeys.posts],
@@ -25,6 +26,8 @@ const usePosts = ({ take, lastIdx = -1 }: Props) => {
           lastPage.posts?.length === take
             ? lastPage.posts[lastPage.posts.length - 1].idx
             : null,
+        // ssr
+        initialData,
       }
     );
